+++ src/gpu/executor/HybridInterplayPatternMatch.cuh
@@ -19,15 +19,15 @@ class HybridInterplayPatternMatch : public HybridSeparatePatternMatch {
   }
 
  protected:
-  virtual void CPUExecute(long long *cpu_ans) {
+  virtual void CPUExecute(size_t *cpu_ans) {
     // CPU does not process inter-partition workload
   }
 
-  virtual void GPUExecute(long long *gpu_ans) {
+  virtual void GPUExecute(size_t *gpu_ans) {
     std::cout << "GPU start interplay processing" << std::endl;
     size_t gpu_thread_num = plan_->GetDevPartitionNum();
-    std::vector<long long> thread_intra_part_ans(gpu_thread_num, 0);
-    std::vector<long long> thread_inter_part_ans(gpu_thread_num, 0);
+    std::vector<size_t> thread_intra_part_ans(gpu_thread_num, 0);
+    std::vector<size_t> thread_inter_part_ans(gpu_thread_num, 0);
 #pragma omp parallel num_threads(gpu_thread_num)
     {
       size_t omp_thread_id = ParallelUtils::GetParallelThreadId();
@@ -51,7 +51,7 @@ class HybridInterplayPatternMatch : public HybridSeparatePatternMatch {
       }
     }
 
-    long long intra_part_gpu_ans = 0;
+    size_t intra_part_gpu_ans = 0;
     for (size_t i = 0; i < gpu_thread_num; ++i) {
       *gpu_ans += thread_inter_part_ans[i] + thread_intra_part_ans[i];
       intra_part_gpu_ans += thread_intra_part_ans[i];
@@ -61,7 +61,7 @@ class HybridInterplayPatternMatch : public HybridSeparatePatternMatch {
   }
 
   virtual void GPUThreadInterPart(size_t omp_thread_id, size_t dev_id,
-                                  long long &thread_ans) {
+                                  size_t &thread_ans) {
     InterPartTask inter_part_task;
     while (inter_scheduler_->GetTask(inter_part_task, false)) {
       std::cout << "thread_id=" << omp_thread_id << ",dev_id=" << dev_id
